Actor: Fix Actor permanently turning invisible after opacity set to 0.
This was a confusing behaviour whereby setting an Actor's opacity = 0
would also set visible = false on the next update cycle. The Actor
would then stay invisible even if the opacity was subsequently set to
a value > 0.

The original behaviour was intended as an optimization. Instead of
setting visible = false when opacity === 0, let's check the opacity
everywhere visible was checked.